Resort functions/components for pipeline graphs
To prepare for pipeline graphs, had to reshuffle the functions and components to be more consistent to their role in their files. ENCD-3510
