+++ src/encoded/static/components/graph.js
@@ -1,13 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import moment from 'moment';
 import _ from 'underscore';
 import { Panel } from '../libs/bootstrap/panel';
 import { svgIcon } from '../libs/svg-icons';
 import { BrowserFeat } from './browserfeat';
-import * as globals from './globals';
 import { requestFiles } from './objectutils';
-import { SortTable } from './sorttable';
 
 
 // Zoom slider constants
@@ -15,7 +12,6 @@ const minZoom = 0; // Minimum zoom slider level
 const maxZoom = 100; // Maximum zoom slider level
 const graphWidthMargin = 40; // Margin on horizontal edges of graph SVG
 const graphHeightMargin = 40; // Margin on vertical edges of graph SVG
-const MINIMUM_COALESCE_COUNT = 5; // Minimum number of files in a coalescing group
 
 
 // The JsonGraph object helps build JSON graph objects. Create a new object
@@ -170,473 +166,6 @@ export function GraphException(message, file0, file1) {
 }
 
 
-// Map a QC object to its corresponding two-letter abbreviation for the graph.
-function qcAbbr(qc) {
-    // As we add more QC object types, add to this object.
-    const qcAbbrMap = {
-        BigwigcorrelateQualityMetric: 'BC',
-        BismarkQualityMetric: 'BK',
-        ChipSeqFilterQualityMetric: 'CF',
-        ComplexityXcorrQualityMetric: 'CX',
-        CorrelationQualityMetric: 'CN',
-        CpgCorrelationQualityMetric: 'CC',
-        DuplicatesQualityMetric: 'DS',
-        EdwbamstatsQualityMetric: 'EB',
-        EdwcomparepeaksQualityMetric: 'EP',
-        Encode2ChipSeqQualityMetric: 'EC',
-        FastqcQualityMetric: 'FQ',
-        FilteringQualityMetric: 'FG',
-        GenericQualityMetric: 'GN',
-        HotspotQualityMetric: 'HS',
-        IDRQualityMetric: 'ID',
-        IdrSummaryQualityMetric: 'IS',
-        MadQualityMetric: 'MD',
-        SamtoolsFlagstatsQualityMetric: 'SF',
-        SamtoolsStatsQualityMetric: 'SS',
-        StarQualityMetric: 'SR',
-        TrimmingQualityMetric: 'TG',
-    };
-
-    let abbr = qcAbbrMap[qc['@type'][0]];
-    if (!abbr) {
-        // 'QC' is the generic, unmatched abbreviation if qcAbbrMap doesn't have a match.
-        abbr = 'QC';
-    }
-    return abbr;
-}
-
-
-/**
- * Assembly a graph of files, the QC objects that belong to them, and the steps that connect them.
- * @param {object} dataset - Dataset object containing the files to graph.
- * @param {object} session - Current login session information.
- * @param {string} infoNodeId - Graph node ID of the currently selected node, whether file, step,
- *        or coalesced contributing file.
- * @param {array} files - Files to graph. This is the *real* source of data. Usually retrieved from
- *        a GET request for qualifying files.
- * @param {string} filterAssembly - Assembly to restrict graphed files to, or '' to not filter any
- *        of those.
- * @param {string} filterAnnotation - Annotation to restrict graphed files to, or '' to not filter
- *        any of those.
- * @param {boolean} colorize - True to add CSS classes to colorize the nodes based on their status.
- */
-export function assembleGraph(files, dataset, options) {
-    // Calculate a step ID from a file's derived_from array.
-    function rDerivedFileIds(file) {
-        if (file.derived_from && file.derived_from.length) {
-            return file.derived_from.sort().join();
-        }
-        return '';
-    }
-
-    // Calculate a QC node ID.
-    function rGenQcId(metric, file) {
-        return `qc:${metric['@id'] + file['@id']}`;
-    }
-
-    /**
-     * Generate a string of CSS classes for a file node. Plass the result into a `className` property of a component.
-     *
-     * @param {object-required} file - File we're generating the statuses for.
-     * @param {bool} active - True if the file is active and should be highlighted as such.
-     * @param (bool) colorize - True to colorize the nodes according to their status by adding a CSS class for their status
-     * @param {string} addClasses - CSS classes to add in addition to the ones generated by the file statuses.
-     */
-    function fileCssClassGen(file, active, colorizeNode, addClasses) {
-        let statusClass;
-        if (colorizeNode) {
-            statusClass = file.status.replace(/ /g, '-');
-        }
-        return `pipeline-node-file${active ? ' active' : ''}${colorizeNode ? ` ${statusClass}` : ''}${addClasses ? ` ${addClasses}` : ''}`;
-    }
-
-    const { infoNodeId, selectedAssembly, selectedAnnotation, colorize } = options;
-    const derivedFileIds = _.memoize(rDerivedFileIds, file => file['@id']);
-    const genQcId = _.memoize(rGenQcId, (metric, file) => metric['@id'] + file['@id']);
-
-    // Begin collecting up information about the files from the search result, and gathering their
-    // QC and analysis pipeline information.
-    const allFiles = {}; // All searched files, keyed by file @id
-    let matchingFiles = {}; // All files that match the current assembly/annotation, keyed by file @id
-    const fileQcMetrics = {}; // List of all file QC metrics indexed by file @id
-    const allPipelines = {}; // List of all pipelines indexed by step @id
-    files.forEach((file) => {
-        // allFiles gets all files from search regardless of filtering.
-        allFiles[file['@id']] = file;
-
-        // matchingFiles gets just the files matching the given filtering assembly/annotation.
-        // Note that if all assemblies and annotations are selected, this function isn't called
-        // because no graph gets displayed in that case.
-        if ((file.assembly === selectedAssembly) && ((!file.genome_annotation && !selectedAnnotation) || (file.genome_annotation === selectedAnnotation))) {
-            // Note whether any files have an analysis step
-            const fileAnalysisStep = file.analysis_step_version && file.analysis_step_version.analysis_step;
-            if (!fileAnalysisStep || (file.derived_from && file.derived_from.length)) {
-                // File has no analysis step or derives from other files, so it can be included in
-                // the graph.
-                matchingFiles[file['@id']] = file;
-
-                // Collect any QC info that applies to this file and make it searchable by file
-                // @id.
-                if (file.quality_metrics && file.quality_metrics.length) {
-                    fileQcMetrics[file['@id']] = file.quality_metrics;
-                }
-
-                // Save the pipeline array used for each step used by the file.
-                if (fileAnalysisStep) {
-                    allPipelines[fileAnalysisStep['@id']] = fileAnalysisStep.pipelines;
-                }
-            } // else file has analysis step but no derived from -- can't include in graph.
-        }
-    });
-
-    // Generate a list of file @ids that other files (matching the current assembly and annotation)
-    // derive from (i.e. files referenced in other files' derived_from). allDerivedFroms is keyed
-    // by the derived-from file @id (whether it matches the current assembly and annotation or not)
-    // and has an array of all files that derive from it for its value. So for example:
-    //
-    // allDerivedFroms = {
-    //     /files/<matching accession>: [matching file, matching file],
-    //     /files/<contributing accession>: [matching file, matching file],
-    //     /files/<missing accession>: [matching file, matching file],
-    // }
-    const allDerivedFroms = {};
-    Object.keys(matchingFiles).forEach((matchingFileId) => {
-        const matchingFile = matchingFiles[matchingFileId];
-        if (matchingFile.derived_from && matchingFile.derived_from.length) {
-            matchingFile.derived_from.forEach((derivedFromAtId) => {
-                // Copy reference to allFiles copy of file. Will be undefined for missing and
-                // contributing files.
-                if (allDerivedFroms[derivedFromAtId]) {
-                    // Already saw a file derive from this one, so add the new reference to the end
-                    // of the array of derived-from files.
-                    allDerivedFroms[derivedFromAtId].push(matchingFile);
-                } else {
-                    // Never saw a file derive from this one, so make a new array with a reference
-                    // to it.
-                    allDerivedFroms[derivedFromAtId] = [matchingFile];
-                }
-            });
-        }
-    });
-    // Remember, at this stage allDerivedFroms includes keys for missing files, files not matching
-    // the chosen assembly/annotation, and contributing files.
-
-    // Filter any "island" files out of matchingFiles -- that is, files that derive from no other
-    // files, and no other files derive from it.
-    matchingFiles = (function matchingFilesFunc() {
-        const noIslandFiles = {};
-        Object.keys(matchingFiles).forEach((matchingFileId) => {
-            const matchingFile = matchingFiles[matchingFileId];
-            if ((matchingFile.derived_from && matchingFile.derived_from.length) || allDerivedFroms[matchingFileId]) {
-                // This file either has derived_from set, or other files derive from it. Copy it to
-                // our destination object.
-                noIslandFiles[matchingFileId] = matchingFile;
-            }
-        });
-        return noIslandFiles;
-    }());
-    if (Object.keys(matchingFiles).length === 0) {
-        throw new GraphException('No graph: no file relationships for the selected assembly/annotation');
-    }
-    // At this stage, any files in matchingFiles will be rendered. We just have to figure out what
-    // other files need rendering, like raw sequencing files, contributing files, and derived-from
-    // files that have a non-matching annotation and assembly.
-
-    const allReplicates = {}; // All file's replicates as keys; each key references an array of files
-    Object.keys(matchingFiles).forEach((matchingFileId) => {
-        // If the file is part of a single biological replicate, add it to an array of files, where
-        // the arrays are in an object keyed by their relevant biological replicate number.
-        const matchingFile = matchingFiles[matchingFileId];
-        let replicateNum = (matchingFile.biological_replicates && matchingFile.biological_replicates.length === 1) ? matchingFile.biological_replicates[0] : undefined;
-        if (replicateNum) {
-            if (allReplicates[replicateNum]) {
-                allReplicates[replicateNum].push(matchingFile);
-            } else {
-                allReplicates[replicateNum] = [matchingFile];
-            }
-        }
-
-        // Add each file that a matching file derives from to the replicates.
-        if (matchingFile.derived_from && matchingFile.derived_from.length) {
-            matchingFile.derived_from.forEach((derivedFromAtId) => {
-                const file = allFiles[derivedFromAtId];
-                if (file) {
-                    replicateNum = (file.biological_replicates && file.biological_replicates.length === 1) ? file.biological_replicates[0] : undefined;
-                    if (replicateNum) {
-                        if (allReplicates[replicateNum]) {
-                            allReplicates[replicateNum].push(matchingFile);
-                        } else {
-                            allReplicates[replicateNum] = [matchingFile];
-                        }
-                    }
-                }
-            });
-        }
-    });
-
-    // Make a list of contributing files that matchingFiles files derive from.
-    const usedContributingFiles = {};
-    if (dataset.contributing_files && dataset.contributing_files.length) {
-        dataset.contributing_files.forEach((contributingFileAtId) => {
-            if (contributingFileAtId in allDerivedFroms) {
-                usedContributingFiles[contributingFileAtId] = allDerivedFroms[contributingFileAtId];
-            }
-        });
-    }
-
-    // Go through each used contributing file and set a property within it showing which files
-    // derive from it. We'll need that for coalescing contributing files.
-    const allCoalesced = {};
-    let coalescingGroups = {};
-    if (Object.keys(usedContributingFiles).length) {
-        // Now use the derivedFiles property of every contributing file to group them into potential
-        // coalescing nodes. `coalescingGroups` gets assigned an object keyed by dataset file ids
-        // hashed to a stringified 32-bit integer, and mapped to an array of contributing files they
-        // derive from.
-        coalescingGroups = _(Object.keys(usedContributingFiles)).groupBy((contributingFileAtId) => {
-            const derivedFiles = usedContributingFiles[contributingFileAtId];
-            return globals.hashCode(derivedFiles.map(derivedFile => derivedFile['@id']).join(',')).toString();
-        });
-
-        // Set a `coalescingGroup` property in each contributing file with its coalescing group's hash
-        // value. That'll be important when we add step nodes.
-        const coalescingGroupKeys = Object.keys(coalescingGroups);
-        if (coalescingGroupKeys && coalescingGroupKeys.length) {
-            coalescingGroupKeys.forEach((groupHash) => {
-                const group = coalescingGroups[groupHash];
-                if (group.length >= MINIMUM_COALESCE_COUNT) {
-                    // Number of files in the coalescing group is at least the minimum number of files we
-                    // allow in a coalescing group. Mark every contributing file in the group with the
-                    // group's hash value in a `coalescingGroup` property that step node can connnect to.
-                    group.forEach((contributingFileAtId) => {
-                        allCoalesced[contributingFileAtId] = groupHash;
-
-                        // Remove coalesced files from usedContributingFiles because we don't want
-                        // to render individual files that have been coalesced.
-                        delete usedContributingFiles[contributingFileAtId];
-                    });
-                } else {
-                    // The number of contributing files in a coalescing group isn't above our
-                    // threshold. Don't use this coalescingGroup anymore and just render them the
-                    // same as normal files.
-                    delete coalescingGroups[groupHash];
-                }
-            });
-        }
-    }
-
-    // See if we have any derived_from files that we have no information on, likely because they're
-    // not released and we're not logged in. We'll render them with information-less dummy nodes.
-    const allMissingFiles = [];
-    Object.keys(allDerivedFroms).forEach((derivedFromFileAtId) => {
-        if (!allFiles[derivedFromFileAtId] && !allCoalesced[derivedFromFileAtId]) {
-            // The derived-from file isn't in our dataset file list, nor in coalesced contributing
-            // files. Now see if it's in non-coalesced contributing files.
-            if (!usedContributingFiles[derivedFromFileAtId]) {
-                allMissingFiles.push(derivedFromFileAtId);
-            }
-        }
-    });
-
-    // Create an empty graph architecture that we fill in next.
-    const jsonGraph = new JsonGraph(dataset.accession);
-
-    // Create nodes for the replicates.
-    Object.keys(allReplicates).forEach((replicateNum) => {
-        if (allReplicates[replicateNum] && allReplicates[replicateNum].length) {
-            jsonGraph.addNode(`rep:${replicateNum}`, `Replicate ${replicateNum}`, {
-                cssClass: 'pipeline-replicate',
-                type: 'Rep',
-                shape: 'rect',
-                cornerRadius: 0,
-            });
-        }
-    });
-
-    // Go through each file matching the currently selected assembly/annotation and add it to our
-    // graph.
-    Object.keys(matchingFiles).forEach((fileId) => {
-        const file = matchingFiles[fileId];
-        const fileNodeId = `file:${file['@id']}`;
-        const fileNodeLabel = `${file.title} (${file.output_type})`;
-        const fileCssClass = fileCssClassGen(file, infoNodeId === fileNodeId, colorize);
-        const fileRef = file;
-        const replicateNode = (file.biological_replicates && file.biological_replicates.length === 1) ? jsonGraph.getNode(`rep:${file.biological_replicates[0]}`) : null;
-        let metricsInfo;
-
-        // Add QC metrics info from the file to the list to generate the nodes later.
-        if (fileQcMetrics[fileId] && fileQcMetrics[fileId].length) {
-            const sortedMetrics = fileQcMetrics[fileId].sort((a, b) => (a['@type'][0] > b['@type'][0] ? 1 : (a['@type'][0] < b['@type'][0] ? -1 : 0)));
-            metricsInfo = sortedMetrics.map((metric) => {
-                const qcId = genQcId(metric, file);
-                return {
-                    id: qcId,
-                    label: qcAbbr(metric),
-                    '@type': ['QualityMetric'],
-                    class: `pipeline-node-qc-metric${infoNodeId === qcId ? ' active' : ''}`,
-                    tooltip: true,
-                    ref: metric,
-                    parent: file,
-                };
-            });
-        }
-
-        // Add a node for a regular searched file.
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-            parentNode: replicateNode,
-            ref: fileRef,
-        }, metricsInfo);
-
-        // Figure out the analysis step we need to render between the node we just rendered and its
-        // derived_from.
-        let stepId;
-        let label;
-        let pipelineInfo;
-        let error;
-        const fileAnalysisStep = file.analysis_step_version && file.analysis_step_version.analysis_step;
-        if (fileAnalysisStep) {
-            // Make an ID and label for the step
-            stepId = `step:${derivedFileIds(file) + fileAnalysisStep['@id']}`;
-            label = fileAnalysisStep.analysis_step_types;
-            pipelineInfo = allPipelines[fileAnalysisStep['@id']];
-            error = false;
-        } else if (derivedFileIds(file)) {
-            // File derives from others, but no analysis step; make dummy step.
-            stepId = `error:${derivedFileIds(file)}`;
-            label = 'Software unknown';
-            pipelineInfo = null;
-            error = true;
-        } else {
-            // No analysis step and no derived_from; don't add a step.
-            stepId = '';
-        }
-
-        // If we have a step to render, do that here.
-        if (stepId) {
-            // Add the step to the graph only if we haven't for this derived-from set already
-            if (!jsonGraph.getNode(stepId)) {
-                jsonGraph.addNode(stepId, label, {
-                    cssClass: `pipeline-node-analysis-step${(infoNodeId === stepId ? ' active' : '') + (error ? ' error' : '')}`,
-                    type: 'Step',
-                    shape: 'rect',
-                    cornerRadius: 4,
-                    parentNode: replicateNode,
-                    ref: fileAnalysisStep,
-                    pipelines: pipelineInfo,
-                    fileId: file['@id'],
-                    fileAccession: file.title,
-                    stepVersion: file.analysis_step_version,
-                });
-            }
-
-            // Connect the file to the step, and the step to the derived_from files
-            jsonGraph.addEdge(stepId, fileNodeId);
-            file.derived_from.forEach((derivedFromAtId) => {
-                const derivedFromFile = allFiles[derivedFromAtId] || allMissingFiles.some(missingFileId => missingFileId === derivedFromAtId);
-                if (derivedFromFile) {
-                    // Not derived from a contributing file; just add edges normally.
-                    const derivedFileId = `file:${derivedFromAtId}`;
-                    if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                        jsonGraph.addEdge(derivedFileId, stepId);
-                    }
-                } else {
-                    // File derived from a contributing file; add edges to a coalesced node
-                    // that we'll add to the graph later.
-                    const coalescedContributing = allCoalesced[derivedFromAtId];
-                    if (coalescedContributing) {
-                        // Rendering a coalesced contributing file.
-                        const derivedFileId = `coalesced:${coalescedContributing}`;
-                        if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                            jsonGraph.addEdge(derivedFileId, stepId);
-                        }
-                    } else if (usedContributingFiles[derivedFromAtId]) {
-                        const derivedFileId = `file:${derivedFromAtId}`;
-                        if (!jsonGraph.getEdge(derivedFileId, stepId)) {
-                            jsonGraph.addEdge(derivedFileId, stepId);
-                        }
-                    }
-                }
-            });
-        }
-    });
-
-    // Go through each derived-from file and add it to our graph.
-    Object.keys(allDerivedFroms).forEach((fileId) => {
-        const file = allFiles[fileId];
-        if (file && !matchingFiles[fileId]) {
-            const fileNodeId = `file:${file['@id']}`;
-            const fileNodeLabel = `${file.title} (${file.output_type})`;
-            const fileCssClass = fileCssClassGen(file, infoNodeId === fileNodeId, colorize);
-            const fileRef = file;
-            const replicateNode = (file.biological_replicates && file.biological_replicates.length === 1) ? jsonGraph.getNode(`rep:${file.biological_replicates[0]}`) : null;
-
-            jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-                cssClass: fileCssClass,
-                type: 'File',
-                shape: 'rect',
-                cornerRadius: 16,
-                parentNode: replicateNode,
-                ref: fileRef,
-            });
-        }
-    });
-
-    // Go through each derived-from contributing file and add it to our graph.
-    Object.keys(usedContributingFiles).forEach((fileAtId) => {
-        const fileNodeId = `file:${fileAtId}`;
-        const fileNodeLabel = `${globals.atIdToAccession(fileAtId)}`;
-        const fileCssClass = `pipeline-node-file contributing${infoNodeId === fileNodeId ? ' active' : ''}`;
-
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-            contributing: fileAtId,
-            ref: {},
-        });
-    });
-
-    // Now add coalesced nodes to the graph.
-    Object.keys(coalescingGroups).forEach((groupHash) => {
-        const coalescingGroup = coalescingGroups[groupHash];
-        if (coalescingGroup.length) {
-            const fileNodeId = `coalesced:${groupHash}`;
-            const fileCssClass = `pipeline-node-file contributing${infoNodeId === fileNodeId ? ' active' : ''}`;
-            jsonGraph.addNode(fileNodeId, `${coalescingGroup.length} contributing files`, {
-                cssClass: fileCssClass,
-                type: 'Coalesced',
-                shape: 'stack',
-                cornerRadius: 16,
-                contributing: groupHash,
-                ref: coalescingGroup,
-            });
-        }
-    });
-
-    // Add missing-file nodes to the graph.
-    allMissingFiles.forEach((missingFileAtId) => {
-        const fileNodeAccession = globals.atIdToAccession(missingFileAtId);
-        const fileNodeId = `file:${missingFileAtId}`;
-        const fileNodeLabel = `${fileNodeAccession} (unknown)`;
-        const fileCssClass = 'pipeline-node-file error';
-
-        jsonGraph.addNode(fileNodeId, fileNodeLabel, {
-            cssClass: fileCssClass,
-            type: 'File',
-            shape: 'rect',
-            cornerRadius: 16,
-        });
-    });
-
-    return jsonGraph;
-}
-
-
 export class Graph extends React.Component {
     // Take a JsonGraph object and convert it to an SVG graph with the Dagre-D3 library.
     // jsonGraph: JsonGraph object containing nodes and edges.
@@ -714,7 +243,7 @@ export class Graph extends React.Component {
     }
 
     componentDidMount() {
-        if (this.cv.graph) {
+        if (this.props.graph) {
             if (BrowserFeat.getBrowserCaps('svg')) {
                 // Delay loading dagre for Jest testing compatibility;
                 // Both D3 and Jest have their own conflicting JSDOM instances
@@ -843,7 +372,7 @@ export class Graph extends React.Component {
         const render = new dagreD3.render();
 
         // Convert from given node architecture to the dagre nodes and edges
-        Graph.convertGraph(this.cv.graph, g);
+        Graph.convertGraph(this.props.graph, g);
 
         // Run the renderer. This is what draws the final graph.
         render(svg, g);
@@ -880,11 +409,11 @@ export class Graph extends React.Component {
         if (nodeId) {
             if (nodeId.indexOf('qc:') >= 0) {
                 // QC subnode.
-                node = this.cv.graph.getSubnode(nodeId);
+                node = this.props.graph.getSubnode(nodeId);
                 node.schemas = this.props.schemas;
             } else if (nodeId.indexOf('coalesced:') >= 0) {
                 // Coalesced contributing files.
-                const coalescedNode = this.cv.graph.getNode(nodeId);
+                const coalescedNode = this.props.graph.getNode(nodeId);
                 if (coalescedNode) {
                     const currCoalescedFiles = this.state.coalescedFiles;
                     if (currCoalescedFiles[coalescedNode.metadata.contributing]) {
@@ -911,7 +440,7 @@ export class Graph extends React.Component {
                 }
             } else {
                 // A regular or contributing file.
-                node = this.cv.graph.getNode(nodeId);
+                node = this.props.graph.getNode(nodeId);
                 if (node) {
                     node.schemas = this.props.schemas;
                     if (node.metadata.contributing) {
@@ -1079,30 +608,12 @@ export class Graph extends React.Component {
     }
 
     render() {
-        const { dataset, files, selectedAssembly, selectedAnnotation } = this.props;
+        const { graph } = this.props;
         const orientBtnAlt = `Orient graph ${this.state.verticalGraph ? 'horizontally' : 'vertically'}`;
         const currOrientKey = this.state.verticalGraph ? 'orientH' : 'orientV';
         const noDefaultClasses = this.props.noDefaultClasses;
 
-        // Build node graph of the files and analysis steps with this experiment
-        if (files.length) {
-            try {
-                this.cv.graph = assembleGraph(
-                    files,
-                    dataset,
-                    {
-                        infoNodeId: this.state.infoNodeId,
-                        selectedAssembly,
-                        selectedAnnotation,
-                        colorize: this.state.colorize,
-                    }
-                );
-            } catch (e) {
-                console.warn(e.message + (e.file0 ? ` -- file0:${e.file0}` : '') + (e.file1 ? ` -- file1:${e.file1}` : ''));
-            }
-        }
-
-        if (this.cv.graph) {
+        if (graph) {
             return (
                 <Panel noDefaultClasses={noDefaultClasses}>
                     <div className="zoom-control-area">
@@ -1131,15 +642,10 @@ export class Graph extends React.Component {
 }
 
 Graph.propTypes = {
-    files: PropTypes.array.isRequired, // Array of files to render
-    dataset: PropTypes.object.isRequired, // Dataset object the graph is being rendered into
-    selectedAssembly: PropTypes.string, // Selected assembly to display
-    selectedAnnotation: PropTypes.string, // Selected annotation to display
+    graph: PropTypes.object, // JsonGraph object representing the graph being rendered.
     nodeClickHandler: PropTypes.func.isRequired, // Function to call to handle clicks in a node
     noDefaultClasses: PropTypes.bool, // True to supress default CSS classes on <Panel> components
     schemas: PropTypes.object, // Schemas for QC metrics
-    auditIndicators: PropTypes.func, // Inherited from auditDecor HOC
-    auditDetail: PropTypes.func, // Inherited from auditDecor HOC
     children: PropTypes.node,
 };
 
@@ -1156,57 +662,4 @@ Graph.defaultProps = {
 Graph.contextTypes = {
     session: PropTypes.object,
     session_properties: PropTypes.object,
-};
-
-
-export class FileGraph extends React.Component {
-    constructor() {
-        super();
-
-        // Initialize React state variables.
-        this.state = {
-            coalescedFiles: {}, // List of coalesced files we've requested; acts as a cache too
-            infoModalOpen: false, // Graph information modal open
-        };
-    }
-
-    render() {
-        const { files, dataset, selectedAssembly, selectedAnnotation, handleNodeClick, schemas } = this.props;
-
-        // Build node graph of the files and analysis steps with this experiment
-        if (files && files.length && (selectedAssembly || selectedAnnotation)) {
-            return (
-                <Graph
-                    files={files}
-                    dataset={dataset}
-                    nodeClickHandler={handleNodeClick}
-                    selectedAssembly={selectedAssembly}
-                    selectedAnnotation={selectedAnnotation}
-                    schemas={schemas}
-                    auditIndicators={this.props.auditIndicators}
-                    auditDetail={this.props.auditDetail}
-                />
-            );
-        }
-        return <p className="browser-error">Graph not applicable.</p>;
-    }
-}
-
-FileGraph.propTypes = {
-    files: PropTypes.array.isRequired, // Array of files we're graphing
-    dataset: PropTypes.object.isRequired, // dataset these files are being rendered into
-    selectedAssembly: PropTypes.string, // Currently selected assembly
-    selectedAnnotation: PropTypes.string, // Currently selected annotation
-    schemas: PropTypes.object, // Schemas for QC metrics
-    handleNodeClick: PropTypes.func.isRequired, // Parent function to call when a graph node is clicked
-    auditIndicators: PropTypes.func, // Inherited from auditDecor HOC
-    auditDetail: PropTypes.func, // Inherited from auditDecor HOC
-};
-
-FileGraph.defaultProps = {
-    selectedAssembly: '',
-    selectedAnnotation: '',
-    schemas: null,
-    auditIndicators: null,
-    auditDetail: null,
 };
